Add FileField for Files.
